pyproject.toml
[tool.poetry]
name = "lazy-type-hint"
version = "0.1.0"
description = ""
authors = ["Manuel Floriano Vázquez <mflovaa@gmail.com>"]
readme = "README.md"
packages = [{include = "lazy_type_hint"}]

[tool.poetry.dependencies]
python = "^3.8, <=3.12"
typing_extensions = "*"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
ruff = "^0.3.2"
pytest = "7.4.4"  # Testing
codespell = "2.2.6"  # Spell checker
black = "23.12.1"  # Autoformatter
pytest-clarity = "^1.0.1"
pyyaml = "*"  # Used by the `check_code.py` script
types-pyyaml = "*"
filelock = "*"
pytest_xdist = "*"  # Parallel execution tests

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Config for developer tools
[tool.black]
line-length = 120
target-version = ["py38"]
include = 'lazy_type_hint/.*\.py$|tests/.*\.py$|check_code\.py$'
force_exclude = """test_files/"""
exclude = """test_files/"""


[tool.mypy]
ignore_missing_imports = true
exclude = ["build", "test_files"]
python_version = "3.8"
strict = true
explicit_package_bases = true

[tool.ruff]
line-length = 120
target-version = "py38"
exclude = ["*.pyi", "test_files"]
lint.select = ["F", "E", "I", "N", "D", "UP", "YTT", "ASYNC", "BLE", "B", "A", "C4", "T10", "DJ", "EXE", "ISC", "ICN", "G", "PIE", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "INT", "ARG", "FIX", "PD", "PGH", "PL", "FLY", "NPY", "AIR", "RUF", "TCH"]
lint.ignore = ["SIM102", "EM101", "PLR", "D100", "D103", "C419", "D104", "D105", "D212", "PGH003", "RUF001", "D101", "PD901", "C405", "UP015", "PLC0414", "I001", "NPY002", "D102", "D107", "PT023", "B027", "UP007", "PIE808", "PD011", "SIM115", "PT006", "PT007", "F811", "UP006", "RET505", "PLW1510", "E731", "PLW2901", "SLF001"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # Docstring convention. Accepts: "google", "numpy", or "pep257".

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]
"tests/**/test_func_data_type_tree.py" = ["ARG005", "ARG001"]

[tool.codespell]
ignore-words-list = "bu"  # Comma separated

[tool.pytest.ini_options]
addopts = "-n auto"
testpaths = ["tests"]